gui: uint64_t -> uint32_t
the psp is a 32 bit system it would be better to use 32 bit ints

+++ src/gui/Animation.h
@@ -45,7 +45,7 @@ struct Ref
 	TweenSpec spec {};
 	std::function<bool()> pause_if {};
 	float fallback {};
-	uint64_t generation {};
+	uint32_t generation {};
 
 	auto valid() const -> bool { return !key.empty(); }
 };
